refactor: 이후 필터의 예외처리를 위한 공통 커스텀 예외 객체 생성 및 패키지 구조 변경
- 이후에 다형성을 이용하여 중복 코드를 줄이기 위해 FilterExceptionHandler에서 이를 적용할 예정입니다.
