+++ src/main/java/kr/adapterz/community/common/exception/dto/ForbiddenException.java
@@ -1,19 +1,20 @@
-package kr.adapterz.community.common.exception;
+package kr.adapterz.community.common.exception.dto;
 
 import kr.adapterz.community.common.message.ErrorCode;
 import org.springframework.http.HttpStatus;
 
-public class NotFoundException extends RuntimeException implements ErrorStatusCodeProvider {
+public class ForbiddenException extends RuntimeException implements ErrorStatusCodeProvider {
+
     /**
-     * HTTP status code 404을 나타내는 예외입니다.
+     * HTTP status code 403을 나타내는 예외입니다.
      */
-    public NotFoundException(ErrorCode message) {
+    public ForbiddenException(ErrorCode message) {
         super(message.getMessage());
     }
 
     @Override
     public HttpStatus getHttpStatus() {
-        return HttpStatus.NOT_FOUND;
+        return HttpStatus.FORBIDDEN;
     }
 
     @Override
