+++ src/main/java/kr/adapterz/community/common/exception/dto/NotFoundException.java
@@ -1,20 +1,19 @@
-package kr.adapterz.community.common.exception;
+package kr.adapterz.community.common.exception.dto;
 
 import kr.adapterz.community.common.message.ErrorCode;
 import org.springframework.http.HttpStatus;
 
-public class ForbiddenException extends RuntimeException implements ErrorStatusCodeProvider {
-
+public class NotFoundException extends RuntimeException implements ErrorStatusCodeProvider {
     /**
-     * HTTP status code 403을 나타내는 예외입니다.
+     * HTTP status code 404을 나타내는 예외입니다.
      */
-    public ForbiddenException(ErrorCode message) {
+    public NotFoundException(ErrorCode message) {
         super(message.getMessage());
     }
 
     @Override
     public HttpStatus getHttpStatus() {
-        return HttpStatus.FORBIDDEN;
+        return HttpStatus.NOT_FOUND;
     }
 
     @Override
